requirements.txt
Django==1.8.3
Pillow==2.9.0
dj-database-url==0.3.0
djangorestframework==3.2.2
gunicorn==19.3.0
whitenoise==2.0.3
